generalised to random forests, minor fix
Added some code to make `importance_table` generalise to random
forests, and also made the if(class) part handle objects that have more
than one class.

In the future I need to work on making all of this into S3, but it
works as is right now, so I don’t really see the advantage in using it.

Hopefully once I’ve worked on BCARTs, BARTs, etc., I can write similar
methods for each of these.