new function ggplot_missing
I received a weird error when I checked neato, which was
```
* checking package namespace information ... OK
* checking package dependencies ... ERROR
Namespace dependency not required: ‘gbm’

See section ‘The DESCRIPTION file’ in the ‘Writing R Extensions’
manual.
* DONE
Status: 1 ERROR
```
I think it has to do with the fact that I have an S3 method with .gbm -
meaning it imports gbm, but it doesn’t need to. Or maybe I have asked
it to import gbm, but it doesn’t need to.